TAQ Loader Process Additions (#9)
* check if file exists and error trap date assignment

* add indent within if statement

* add loadstatus to return dictionary

* add comments

* return error if load fails

* update loadstatus if load is successful

* exit loader with error if file not found or if date not extracted

* fix error trap when .Q.fpn is called

* Mergework (#8)

* initial qmerger

* WIP, not yet usable

* WIP mark 2

* works but no error trapping yet

* completed merger function

* Remove testing steps

* introduce environment variables

* add move to hdb function

* move test lines to bottom

* set tempdb environment variable

* Change merge to output boolean

* add log messages in merge

* add log messages in reset

* key -merged- dictionary by date and split

* remove split file loading and unenumerating

* move to hdb using system commands

* uncomment logging

* make global split variable local

* move setenv.sh

* persist merged table

* make date a variable

* set merged table as environment variable

* change TORQMERGED env variable to TORQTAQMERGED

* change in setenv.sh

* remove global path assignments

* remove test lines

* remove global split variable

* add processes to process.csv

* add config for qmerger

* add default values to qmerger

* comment file alerter for testing

* move clear tempdb command

* remove unnecessary lines

* remove homedir

* update pardir and quotedir

* fix TORQTAQMERGED environment variable

* git remove / from tempdb env variable

* merge to /path/to/tempdb/final/date/quote

* move syscmd function to common

* comment explaining no.17 in line 30

* fix merge conflict

* last commit

* remove homedir from qmerger settings

* remove magic numbers


* handle when taqloader is not available

* add extra param to loadtaqfile call in runload

* check if file exists and error trap date assignment

* add indent within if statement

* add loadstatus to return dictionary

* add comments

* return error if load fails

* update loadstatus if load is successful

* exit loader with error if file not found or if date not extracted

* fix error trap when .Q.fpn is called

* handle when taqloader is not available

* add extra param to loadtaqfile call in runload

* fix filepath in loadtaqfile

* add loadstatus to fileloading schema

* remove extra bracket

+++ code/processes/taqloader.q
@@ -55,40 +55,52 @@ nbboparams:defaults,(!) . flip (
         );
 
 // function to load all taq files from nyse
-loadtaqfile:{[filetype;filetoload;loadid;optionalparams]
-  filepath:raze (getenv[`TORQTAQFILEDROP]),string filetoload;
-  // define params based on filetype
-  params:$[
-    filetype=`trade;tradeparams,optionalparams;
-    filetype=`quote;quoteparams,optionalparams;
-    filetype=`nbbo;nbboparams,optionalparams;
-    [.lg.e[`fifoloader;errmsg:(string filetype)," is an unknown or unsupported file type"];'errmsg]
-    ];
-  // if quote then partition by letter in the temp hdb
-  params[`dbdir]:$[
-    filetype=`trade;`$(string params[`tempdb]),"/",(string filetype);
-    filetype=`quote;`$(string params[`tempdb]),"/",(string filetype),last -12_string filetoload;
-    `$(string params[`tempdb]),"/",(string filetype);
-    ];
-  // make fifo with PID attached
-  fifo:"/tmp/fifo",string .z.i;
-  // extract date
-  date:"D"$-8#-3_string filetoload;
-  params[`date]:date;
-  // remove fifo if it exists then make new one
-  syscmd["rm -f ",fifo," && mkfifo ",fifo];
-  syscmd["gunzip -c ",(filepath)," > ",fifo," &"];
-  .lg.o[`fifoloader;"Loading ",(string filetoload)];
-  .Q.fpn[.loader.loaddata[params,(enlist`filename)!enlist `$-3_string filetoload];hsym `$fifo;params`chunksize];
-  .lg.o[`fifoloader;(string filetoload)," has successfully been loaded"];
-  syscmd["rm ",fifo];
+loadtaqfile:{[filetype;filetoload;filepath;loadid;optionalparams]
+  foundfile:1b;
+  // initialize as fail and update to success if fully loaded
+  loadstatus:`fail;
+  // hard code numbers in date assignment since file names are uniform
+  date:@[{"D"$-8#-3_string x};filetoload;0Nd];
+  if[0Nd=date;[.lg.e[`loadtaqfile;errmsg:("Could not extract date in "),string filetoload];'errmsg]];
+  $[filetoload in key[hsym`$getenv[`TORQTAQFILEDROP]];
+    .lg.o[`loadtaqfile;raze "File successfully found in ", getenv[`TORQTAQFILEDROP]];
+    foundfile:0b];
+  if[not foundfile;[.lg.e[`loadtaqfile;errmsg:"Could not find ", .os.pth filepath];'errmsg]];  
+  if[foundfile;
+    // define params based on filetype
+    params:$[
+      filetype=`trade;tradeparams,optionalparams;
+      filetype=`quote;quoteparams,optionalparams;
+      filetype=`nbbo;nbboparams,optionalparams;
+      [.lg.e[`fifoloader;errmsg:(string filetype)," is an unknown or unsupported file type"];'errmsg]
+      ];
+    // if quote then partition by letter in the temp hdb
+    params[`dbdir]:$[
+      filetype=`trade;`$(string params[`tempdb]),"/",(string filetype);
+      filetype=`quote;`$(string params[`tempdb]),"/",(string filetype),last -12_string filetoload;
+      `$(string params[`tempdb]),"/",(string filetype);
+      ];
+    // make fifo with PID attached
+    fifo:"/tmp/fifo",string .z.i;
+    params[`date]:date;
+    // remove fifo if it exists then make new one
+    syscmd["rm -f ",fifo," && mkfifo ",fifo];
+    syscmd["gunzip -c ",(.os.pth filepath)," > ",fifo," &"];
+    .lg.o[`fifoloader;"Loading ",(string filetoload)];
+    .[{.Q.fpn[x;y;z]};(.loader.loaddata[params,(enlist`filename)!enlist `$-3_string filetoload];hsym `$fifo;params`chunksize);
+      {[e] [.lg.e[`loadtaqfile;errmsg:"Failed to load file with error:",e];'errmsg]}];
+    .lg.o[`fifoloader;(string filetoload)," has successfully been loaded"];
+    syscmd["rm ",fifo];
+    loadstatus:`success;
+  ];
   // result to send to postback function to orchestrator
   (!) . flip (
     (`tablepath;hsym`$(string params[`dbdir]),"/",(string date),"/",(string filetype));
     (`tabletype;filetype);
     (`loadid;loadid);
     (`tabledate;date);
-    (`loadendtime;.z.P)
+    (`loadendtime;.proc.cp[]);
+    (`loadstatus;loadstatus)
   )
  };
